fix: small clean code fixes
