+++ src/routes/graphql/queries/memberType.ts
@@ -1,7 +1,7 @@
 import { GraphQLFieldConfig, GraphQLList, GraphQLNonNull } from 'graphql';
-import { MemberTypeIdType } from '../types/member-id.type.js';
-import { MemberTypeType } from '../types/member.type.js';
-import { Context } from '../types/context.js';
+import { Context } from '../gqlContext.js';
+import { MemberTypeType } from '../types/memberType.js';
+import { MemberTypeIdType } from '../types/memberTypeId.js';
 
 interface Args {
   id: string;
@@ -10,11 +10,11 @@ interface Args {
 export const memberType: GraphQLFieldConfig<void, Context, Args> = {
   type: MemberTypeType,
   args: { id: { type: new GraphQLNonNull(MemberTypeIdType) } },
-  resolve: (_source, args, ctx) =>
-    ctx.prisma.memberType.findUnique({ where: { id: args.id } }),
+  resolve: (_source, args, context) =>
+    context.prisma.memberType.findUnique({ where: { id: args.id } }),
 };
 
 export const memberTypes: GraphQLFieldConfig<void, Context, void> = {
   type: new GraphQLNonNull(new GraphQLList(MemberTypeType)),
-  resolve: (_src, _, ctx) => ctx.prisma.memberType.findMany(),
+  resolve: (_src, _, context) => context.prisma.memberType.findMany(),
 };
