+++ src/routes/graphql/queries/user.ts
@@ -4,9 +4,9 @@ import {
   GraphQLNonNull,
   GraphQLObjectType,
 } from 'graphql';
-import { UserType } from '../types/user.type.js';
-import { UUIDType } from '../types/uuid.type.js';
-import { Context } from '../types/context.js';
+import { Context } from '../gqlContext.js';
+import { UUIDType } from '../types/uuid.js';
+import { UserType } from '../types/user.js';
 
 interface Args {
   id: string;
@@ -15,10 +15,11 @@ interface Args {
 export const user: GraphQLFieldConfig<void, Context, Args> = {
   type: UserType as GraphQLObjectType<void, Context>,
   args: { id: { type: new GraphQLNonNull(UUIDType) } },
-  resolve: (_source, { id }, ctx) => ctx.prisma.user.findUnique({ where: { id } }),
+  resolve: (_source, { id }, context) =>
+    context.prisma.user.findUnique({ where: { id } }),
 };
 
 export const users: GraphQLFieldConfig<void, Context, void> = {
   type: new GraphQLNonNull(new GraphQLList(UserType)),
-  resolve: (_source, _args, ctx) => ctx.prisma.user.findMany(),
+  resolve: (_source, _args, context) => context.prisma.user.findMany(),
 };
