--- src/routes/graphql/types/context.ts
@@ -1,5 +1,0 @@
-import { PrismaClient } from '@prisma/client';
-
-export interface Context {
-  prisma: PrismaClient;
-}
